requirements.txt
# Core Dependencies
python-dotenv==1.0.1

# LLM & Chat Framework
langchain==0.2.0
langchain-core==0.2.0
langchain-openai==0.2.0
langgraph==0.2.0

# Vector Database
qdrant-client==1.14.0

# Embedding Models
torch>=2.0.0            # CUDA 12.8 compatible
sentence-transformers>=2.2.0
FlagEmbedding>=1.3.5    # For BAAI/bge-m3
transformers>=4.41.0

# PDF Processing
pdfplumber>=0.10.0
docling>=1.0.0          # For OCR + Markdown conversion

# HTTP & Data Processing
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tqdm>=4.60.0
